Updated the fix to keep up with the code change.
Now we have BlocKEndNode inserted, so how we get to AtomNode for the shell step needs to be tweaked.

Also, I believe the test code has a race condition; after shell script checker has run, we need to let the workflow
to "return" from the shell method and run to the completion.