Note to myself.
The test fails right now because CpsStepContext is @PersistIn(ANYWHERE), meaning in general there can be multiple copies of the same instance just because it's serialized multiple times to different places.

So it can't maintain state like #outcome, #startedBodies, and so on. Such stuff would have to be maintained elsewhere.