Allow CPS global lib scripts to define global variables.
This allows the release engineering team to define higher-level
organization specific workflow DSLs.

This is critical to use workflow as "simplified text-based UX".
Such DSLs could be for example:

acmeCorp {
    id = '123'
    devEnvironment = 'qa-svr-1'
}

(where let's say there's an inventory of application ID in the
central database describing where the repository is, and 'qa-svr-1'
refers to the Tomcat server where the app gets deployed at the end.)