Use a modified version of CpsClosure to make the 'sleep' method work.
The method selection logic in the body of a closure is such that a method
defined on groovy.lang.Closure object takes a precedence over any methods
exposed via 'owner' or 'delegate'. This applies to so-called expando methods
that are added externally (such as sleep), not just methods inheretently
defined by JDK (such as hashCode). The following code illustrates this behavior:

=====
class Foo {
    public void test() {
        sleep(3); // this prints fake-sleep(3)
        def c = { sleep(4); } // this does NOT print fake-sleep(4)
        c();
    }

    public void sleep(long l) {
        System.out.println("fake-sleep(${l})");
    }
}

new Foo().test()
=====

Before PR 131, CPS transformation was incorrectly converting 'sleep(4)' above
to 'this.sleep(4)', which will print out fake-sleep(4) because 'this' points
to the Foo instance and not the closure instance.
In PR 131, I fixed this bug in groovy-cps, which means now the sleep()
call in the workflow is not calling into SleepStep, and instead rather
goes into 'java.lang.Object.sleep(long)' added by Groovy.

This change fixes it, in conjunction with cloudbees/groovy-sandbox #13,
by defining the sleep() method explicitly on the closure class we use
(CpsClosure2), and delegate that call to the closure 'owner'.

The net result is that in workflow, such call will get routed to the
DSL.invokeMethod() which then goes to SleepStep.

See: https://github.com/cloudbees/groovy-cps/pull/13