Adding the pause functionality
@uaarkoti told me that he talks to lots of people who wants to be able
to arbitrarily pause workflows while they are executing.

This seems like a relatively easy addition as far as the engine is
concerned, so this is a prototype of that functionality.

Given the way the things are split into plugins I'm not sure how to
properly expose this to UI --- should the pausability be a base contract
of FlowExecution, or should I define TransientActionFactory impl that
depends on both WorkflowRun and CpsFlowExecution that adds an Action
from outside?